io: Renamed src/file to src/io
Fixes #11980.
