+++ Android.mk
@@ -35,8 +35,8 @@ LOCAL_SRC_FILES := \
 	$(LOCAL_PATH)/src/dialog/android/SDL_androiddialog.c \
 	$(wildcard $(LOCAL_PATH)/src/dynapi/*.c) \
 	$(wildcard $(LOCAL_PATH)/src/events/*.c) \
-	$(wildcard $(LOCAL_PATH)/src/file/*.c) \
-	$(wildcard $(LOCAL_PATH)/src/file/generic/*.c) \
+	$(wildcard $(LOCAL_PATH)/src/io/*.c) \
+	$(wildcard $(LOCAL_PATH)/src/io/generic/*.c) \
 	$(wildcard $(LOCAL_PATH)/src/gpu/*.c) \
 	$(wildcard $(LOCAL_PATH)/src/gpu/vulkan/*.c) \
 	$(wildcard $(LOCAL_PATH)/src/haptic/*.c) \
