+++ src/process/posix/SDL_posixprocess.c
@@ -34,7 +34,7 @@
 #include <sys/wait.h>
 
 #include "../SDL_sysprocess.h"
-#include "../../file/SDL_iostream_c.h"
+#include "../../io/SDL_iostream_c.h"
 
 
 #define READ_END 0
